refactor(ecies): rename proto to stream
