+++ crates/net/ecies/src/stream.rs
